+++ iotdb-core/node-commons/pom.xml
@@ -52,11 +52,6 @@
         </dependency>
         <dependency>
             <groupId>org.apache.iotdb</groupId>
-            <artifactId>iotdb-session</artifactId>
-            <version>2.0.0-SNAPSHOT</version>
-        </dependency>
-        <dependency>
-            <groupId>org.apache.iotdb</groupId>
             <artifactId>trigger-api</artifactId>
             <version>2.0.0-SNAPSHOT</version>
         </dependency>
